requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=1.5.0
scipy>=1.10.0
matplotlib>=3.6.0
opencv-python>=4.7.0
Pillow>=9.4.0
seaborn>=0.12.0
scikit-learn>=1.2.0

